moved redis-logic to seperate package
