+++ backend/main.go
@@ -3,6 +3,7 @@ package main
 import (
 	"Omiran-App/backend/dbutils"
 	"Omiran-App/backend/handler"
+	"Omiran-App/backend/redisutils"
 
 	"github.com/gin-gonic/gin"
 )
@@ -10,17 +11,17 @@ import (
 func init() {
 	dbutils.Open("database-config.yaml")
 	handler.InitGQLSchema()
+	redisutils.InitCache()
 }
 
 func main() {
 	r := gin.Default()
-	handler.InitCache()
 
 	r.POST("/graphql", handler.GraphQLService)
 	r.POST("/create", handler.AccountCreationHandler)
 
 	r.POST("/auth", handler.AuthHandler)
 	r.POST("/follow", handler.StartFollowingHandler)
-	handler.Examine()
+
 	r.Run()
 }
