+++ backend/redisutils/redisutils.go
@@ -1,0 +1,16 @@
+package redisutils
+
+import (
+	"github.com/gomodule/redigo/redis"
+)
+
+var cache redis.Conn
+
+//InitCache creates the cache
+func InitCache() {
+	conn, err := redis.DialURL("redis://localhost")
+	if err != nil {
+		panic(err)
+	}
+	cache = conn
+}
